size_t casting

+++ 101-binary_tree_levelorder.c
@@ -11,7 +11,7 @@
 
 void aux(const binary_tree_t *tree, int i, void (*func)(int))
 {
-	if(binary_tree_depth(tree) == (int) i)
+	if((size_t) binary_tree_depth(tree) == i)
 		func(tree->n);
 	aux(tree->left, i, func);
 	aux(tree->right, i, func);
